feat: myRegisterClubDetailFrame 삭제
